// src/app/shared/Models/GoldRate/gold-rate.model.ts
export class GoldRate
{
    GoldRateId: number;
    RateDate: Date | string | null;
    K12: number | null;
    K13: number | null;
    K14: number | null;
    K15: number | null;
    K16: number | null;
    K17: number | null;
    K18: number | null;
    K19: number | null;
    K20: number | null;
    K21: number | null;
    K22: number | null;
    K23: number | null;
    K24: number | null;

    //
    K12Tola: number | null;
    K13Tola: number | null;
    K14Tola: number | null;
    K15Tola: number | null;
    K16Tola: number | null;
    K17Tola: number | null;
    K18Tola: number | null;
    K19Tola: number | null;
    K20Tola: number | null;
    K21Tola: number | null;
    K22Tola: number | null;
    K23Tola: number | null;
    K24Tola: number | null;
    DolleeounceRate: number | null;
}

export class GoldRates
{
    GoldRateId: number;
    RateDate: Date | string | null;
    K12: number | null;
    K13: number | null;
    K14: number | null;
    K15: number | null;
    K16: number | null;
    K17: number | null;
    K18: number | null;
    K19: number | null;
    K20: number | null;
    K21: number | null;
    K22: number | null;
    K23: number | null;
    K24: number | null;

    //
    K12Tola: number | null;
    K13Tola: number | null;
    K14Tola: number | null;
    K15Tola: number | null;
    K16Tola: number | null;
    K17Tola: number | null;
    K18Tola: number | null;
    K19Tola: number | null;
    K20Tola: number | null;
    K21Tola: number | null;
    K22Tola: number | null;
    K23Tola: number | null;
    K24Tola: number | null;
    DolleeounceRate: number | null;
}